исправления  по коментариям наставника
