+++ js/main.js
@@ -1,10 +1,10 @@
 import './similar-element.js';
 
 
-const popup = document.querySelectorAll('.popup');
+const advertisements = document.querySelectorAll('.popup');
 
 const showAdvertisement = (element, n) => {
   element[n].classList.remove('visually-hidden');
 };
 
-showAdvertisement(popup, 0);
+showAdvertisement(advertisements, 0);
