+++ js/similar-element.js
@@ -1,4 +1,5 @@
 import {getArrayAdvertisements} from './advertisement.js';
+import {Housing} from './datum-initial.js'
 
 const advertisementList = document.querySelector('.map__canvas');
 const templateFragment = document.querySelector('#card')
@@ -76,17 +77,11 @@ similarAdvertisements.forEach(({author, offer}) => {
 
 
   const typeOfHousing = newAdvertisement.querySelector('.popup__type');
-  const Housing = {
-    palace: 'Дворец',
-    flat: 'Квартира',
-    house: 'Дом',
-    bungalow: 'Бунгало',
-  };
 
-  const getTypeHousing = (element, type) => {
+  const setTypeHousing = (element, type) => {
     return element.textContent = Housing[type];
   };
-  getTypeHousing(typeOfHousing, offer.type);
+  setTypeHousing(typeOfHousing, offer.type);
 
 
   similarListFragment.appendChild(newAdvertisement);
